Add files via upload
I made this to make it easier for my co-workers and I to manually delete an AVID (video editor) database file, which causes AVID to rebuild it.  This was necessary for us to do every once in a while because of some lameness with AVID.  We have three AVID machines (1,2, and 4), and each AVID has it's own shared storage space. When that space reaches a certain amount of space used, AVID automatically creates a new space in the same folder, incrementing the file name by 1.  This will help you understand what some of my instructions do.